別ファイルに独立させるのをし忘れていたので対応
